moving things
new tile file for Tile class, moved whatDoYourSnakeEyesSee to parseBoard file, added imports for Tile class to main and parseBoard
